First meaningful commit
Created a bunch of interfaces to interact with Amiibo / NTAG215 stickers.

`$ node nfc-pcsc.js` - will wait for Amiibos to be scanned and print their names to the console

`$node nfc-pcsc-file.js input [output]` - will parse a file on disk as an Amiibo tag and print out the name. If an `output` filename is provided, it will "clone" the input tag into the `output` file path.